refactor: add create setting func
